Set consult-org-roam-buffer-enabled to nil
This makes consult-org-roam-buffer stop grouping org roam buffers in a
dedicated category in consult-buffer. This makes consult-buffer's
behavior consistent when consult-org-roam is not used. The default
candidate of consult-buffer is the last buffer you have visited and
not displayed. If the option is set to true, it makes the org roam
buffers in a separate category in consult-buffer, which breaks the
default behavior when your last visited and not displayed buffer is an
org roam buffer.